Update User API (#101)

+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/service/RefreshService.kt
@@ -20,7 +20,7 @@ class RefreshService(
             return null
         }
 
-        val user = userRepository.findById(id = refreshTokenUserId.toInt()) ?: return null
+        val user = userRepository.findById(id = refreshTokenUserId.toLong()) ?: return null
         return tokenService.createToken(user)
     }
 }
